+++ EventStream/transformer/zero_shot/zero_shot_utils.py
@@ -22,14 +22,14 @@ from torchmetrics.classification import (
     MultilabelAveragePrecision,
 )
 
-from ..data.pytorch_dataset import PytorchDataset
-from ..data.types import PytorchBatch
-from ..utils import task_wrapper
-from .config import StructuredTransformerConfig
-from .model import ESTForGenerativeSequenceModeling
-from .model_output import StreamClassificationModelOutput
-from .stream_classification_lightning import FinetuneConfig
-from .utils import safe_weighted_avg, str_summary
+from ...data.pytorch_dataset import PytorchDataset
+from ...data.types import PytorchBatch
+from ...utils import task_wrapper
+from ..config import StructuredTransformerConfig
+from ..model import ESTForGenerativeSequenceModeling
+from ..model_output import StreamClassificationModelOutput
+from ..stream_classification_lightning import FinetuneConfig
+from ..utils import safe_weighted_avg, str_summary
 from .zero_shot_labeler import Labeler
